Add support for MinGW-w64 * Does not add support for ASSIMP * When catching exceptions from runtime compiled code, the exception address not captured * libfreetype.a compiled without debug info using MinGW-w64 i686-5.3.0-win32-dwarf-rt_v4-rev0 * libfreetypex64.a compiled without debug info using MinGW-w64 x86_64-5.3.0-win32-seh-rt_v4-rev0
Compilers Tested:
* MinGW-w64 i686-4.9.1-win32-dwarf-rt_v3-rev1
* MinGW-w64 i686-5.3.0-win32-dwarf-rt_v4-rev0
* MinGW-w64 x86_64-4.9.1-win32-seh-rt_v3-rev1
* MinGW-w64 x86_64-5.3.0-win32-seh-rt_v4-rev0
* VS 2015 x86 & x86_64
