+++ Aurora/Renderer/AURenMesh.cpp
@@ -19,7 +19,7 @@
 #include "AURenMesh.h"
 
 
-#ifndef _WIN32
+#if !defined(_WIN32) || defined(__MINGW32__)
     #define NO_ASSIMP //Currently not adding assimp support to other platforms
 #endif
