+++ Aurora/Systems/LogSystem/ThreadsafeLogSystem/ThreadsafeLogSystem_PlatformWindows.cpp
@@ -19,7 +19,6 @@
 
 #ifdef _WIN32
     #include <Windows.h>
-    #include <concrt.h>
 #else
 #include <pthread.h>
